Beta

+++ src/main/java/com/metalancer/backend/orders/domain/OrderProducts.java
@@ -4,6 +4,7 @@ import com.metalancer.backend.common.constants.ClaimStatus;
 import com.metalancer.backend.common.constants.ClaimType;
 import com.metalancer.backend.common.constants.OrderStatus;
 import com.metalancer.backend.products.domain.ProductsDetail;
+import java.math.BigDecimal;
 import lombok.Builder;
 import lombok.Getter;
 
@@ -13,13 +14,13 @@ public class OrderProducts {
 
     private ProductsDetail productsDetail;
     private String orderProductNo;
-    private Integer price;
+    private BigDecimal price;
     private OrderStatus orderProductStatus;
     private ClaimType claimType;
     private ClaimStatus claimStatus;
 
     @Builder
-    public OrderProducts(ProductsDetail productsDetail, String orderProductNo, Integer price,
+    public OrderProducts(ProductsDetail productsDetail, String orderProductNo, BigDecimal price,
         OrderStatus orderProductStatus, ClaimType claimType, ClaimStatus claimStatus) {
         this.productsDetail = productsDetail;
         this.orderProductNo = orderProductNo;
